+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Get control characters from termios for emulating SIGINT/SIGQUIT/EOF.
+- Add client:application.name and client:application.process.binary tags.
 
 ## [0.4.0] - 2025-05-18